simplevla/pyproject.toml
[project]
name = "simplevla"
version = "0.1.0"
description = "Simple Vision-Language-Action model from scratch"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "Pillow>=9.5.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
]
